Rename aitemplate folder to ait to resolve import issues
